backend/requirements.txt
fastapi
uvicorn
anthropic
youtube-transcript-api
pydantic
google-generativeai
supabase
python-dotenv